+++ Package.swift
@@ -43,7 +43,6 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/Alamofire/Alamofire.git", "5.2.2")
-        .package(url: "https://github.com/tristanhimmelman/ObjectMapper", "4.2.0")
     ],
     targets: [
         .target(