+++ Source/CDYelpCategoryResponse.swift
@@ -25,16 +25,11 @@
 //  THE SOFTWARE.
 //
 
-import ObjectMapper
+public struct CDYelpCategoryResponse: Codable {
 
-public class CDYelpCategoryResponse: Mappable {
+    public let category: CDYelpCategory?
 
-    public var category: CDYelpCategory?
-
-    public required init?(map: Map) {
-    }
-
-    public func mapping(map: Map) {
-        category    <- map["category"]
+    enum CodingKeys: String, CodingKey {
+        case category
     }
 }